+++ backend/src/baserow/contrib/database/api/views/serializers.py
@@ -19,7 +19,8 @@ class FieldOptionsField(serializers.Field):
         "invalid_value": "Must be valid field options.",
     }
 
-    def __init__(self, serializer_class, **kwargs):
+    def __init__(self, serializer_class, create_if_missing=True, **kwargs):
+        self.create_if_missing = create_if_missing
         self.serializer_class = serializer_class
         self._spectacular_annotation = {
             "field": serializers.DictField(
@@ -77,14 +78,12 @@ class FieldOptionsField(serializers.Field):
         """
 
         if isinstance(value, View):
+            # If the fields are in the context we can pass them into the
+            # `get_field_options` call so that they don't have to be fetched from
+            # the database again.
             field_options = self.context.get("field_options", None)
             if field_options is None:
-                # If the fields are in the context we can pass them into the
-                # `get_field_options` call so that they don't have to be fetched from
-                # the database again.
-                field_options = value.get_field_options(
-                    True, self.context.get("fields")
-                )
+                field_options = value.get_field_options(self.create_if_missing)
             return {
                 field_options.field_id: self.serializer_class(field_options).data
                 for field_options in field_options